+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+Version 0.3
+-------------
+- Added (soft) EM algorithm with GMMs.
+- Added pdf for multivariate gaussian distribution.
+- Added bayesian information criterion (bic).
+- Moved renamed and moved generate_data() (from ann.py) to tools and renamed to xnor.
+
 Version 0.2
 -------------
 - Added first mixture models algorithm (k-means).